+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial support for embeddings/vectors
 - Support for dimensionality reduction
+- Initial implementation of MMD-Critic
 
 ## [0.3.8] - 2021-09-07
 ### Changed